+++ embedded-io/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `core::error::Error` implementations for every custom `impl Error`
 - Migrated `std` feature-gated `std::error::Error` implementations to `core::error::Error`
 - Increased MSRV to 1.81 due to `core::error::Error`
+- Implemented `ReadReady` for `&[u8]` and `WriteReady` for `&mut [u8]`
 
 ## 0.6.1 - 2023-10-22